requirements.txt.txt
python
mysql-connector-python
Pillow
